// src/app/features/categoria/models/resumo.ts
export interface CategoriaResumoResponseDTO {
    totalCategorias: number;
    categoriasAtivas: number;
    categoriaMaisUtilizada: string;
    iconeMaisUtilizada: string;
    corMaisUtilizada: string;
    qtdMaisUtilizada: number;
    categoriaMenosUtilizada: string;
    iconeMenosUtilizada: string;
    corMenosUtilizada: string;
    qtdMenosUtilizada: number;
}